backend/requirements.txt
Flask
python-dotenv
SQLAlchemy
requests 